+++ WebApi/InWords.WebApi/InWords.WebApi/Modules/Blazor/InModuleInitializer.cs
@@ -9,9 +9,10 @@ namespace InWords.WebApi.Modules.Blazor
     {
         public override void ConfigureApp(IApplicationBuilder app)
         {
+            app.UseBlazorFrameworkFiles();
             app.UseEndpoints(endpoints =>
             {
-                endpoints.MapDefaultControllerRoute();
+                //endpoints.MapDefaultControllerRoute();
                 endpoints.MapFallbackToFile("index.html");
             });
         }