+++ WebApi/InWords.WebApi/InWords.WebApi/Modules/DictionaryServiceHandler/Words/GetTrainingWords.cs
@@ -11,12 +11,12 @@ using static InWords.Protobuf.TrainingReply.Types;
 
 namespace InWords.WebApi.Modules.DictionaryServiceHandler.Words
 {
-    public class GetTrainingWords : AuthorizedRequestObjectHandler<Empty, TrainingReply, InWordsDataContext>
+    public class GetTrainingWords : AuthReqHandler<Empty, TrainingReply, InWordsDataContext>
     {
         public GetTrainingWords(InWordsDataContext context) : base(context) { }
 
         public override async Task<TrainingReply> HandleRequest(
-            AuthorizedRequestObject<Empty, TrainingReply> request,
+            AuthReq<Empty, TrainingReply> request,
             CancellationToken cancellationToken = default)
         {
             if (request == null)