+++ WebApi/InWords.WebApi/InWords.WebApi/Modules/Profile/EmailUpdate/ConfirmEmail.cs
@@ -11,14 +11,14 @@ using System.Threading.Tasks;
 
 namespace InWords.WebApi.Services.Users.EmailUpdate
 {
-    public class ConfirmEmail : AuthorizedRequestObjectHandler<ConfirmEmailRequest, ConfirmEmailReply, InWordsDataContext>
+    public class ConfirmEmail : AuthReqHandler<ConfirmEmailRequest, ConfirmEmailReply, InWordsDataContext>
     {
         public ConfirmEmail(InWordsDataContext context) : base(context)
         {
         }
 
         public override async Task<ConfirmEmailReply> HandleRequest(
-            AuthorizedRequestObject<ConfirmEmailRequest, ConfirmEmailReply> request,
+            AuthReq<ConfirmEmailRequest, ConfirmEmailReply> request,
             CancellationToken cancellationToken = default)
         {
             if (request == null)