+++ WebApi/InWords.WebApi/InWords.WebApi/Modules/Profile/PublicData/FindProfileNickname.cs
@@ -12,12 +12,12 @@ using System.Threading.Tasks;
 
 namespace InWords.WebApi.Modules.Profile.PublicData
 {
-    public class FindProfileNickname : AuthorizedRequestObjectHandler<FindUsernameRequest, PublicProfilesReply, InWordsDataContext>
+    public class FindProfileNickname : AuthReqHandler<FindUsernameRequest, PublicProfilesReply, InWordsDataContext>
     {
         public FindProfileNickname(InWordsDataContext context) : base(context) { }
 
         public override Task<PublicProfilesReply> HandleRequest(
-            AuthorizedRequestObject<FindUsernameRequest, PublicProfilesReply> request,
+            AuthReq<FindUsernameRequest, PublicProfilesReply> request,
             CancellationToken cancellationToken = default)
         {
             if (request == null)