+++ WebApi/InWords.WebApi/InWords.WebApi/Modules/WordsSets/GetWordSetsHandler.cs
@@ -1,6 +1,7 @@
 ﻿using InWords.Data;
 using InWords.Data.Enums;
 using InWords.Protobuf;
+using InWords.WebApi.Modules.WordsSets.Extentions;
 using InWords.WebApi.Services.Abstractions;
 using Microsoft.EntityFrameworkCore;
 using System.Linq;
@@ -11,39 +12,15 @@ using static InWords.Protobuf.WordSetReply.Types;
 namespace InWords.WebApi.Modules.WordsSets
 {
     public class GetWordSetsHandler
-        : AuthorizedRequestObjectHandler<Empty, WordSetReply, InWordsDataContext>
+        : AuthReqHandler<Empty, WordSetReply, InWordsDataContext>
     {
         public GetWordSetsHandler(InWordsDataContext context) : base(context)
         {
         }
 
-        public override async Task<WordSetReply> HandleRequest(
-            AuthorizedRequestObject<Empty, WordSetReply> request, CancellationToken cancellationToken = default)
-        {
-            WordSetReply wordSetReply = new WordSetReply();
+        public override Task<WordSetReply> HandleRequest(
+            AuthReq<Empty, WordSetReply> request, CancellationToken cancellationToken = default)
+            => Context.GetWordSets();
 
-            var games = (from tags in Context.GameTags.Where(d => d.Tags == GameTags.Official)
-                         join game in Context.Games on tags.GameId equals game.GameId
-                         select game);
-
-            var include = await games.Include(d => d.CreationDescriptions)
-                .ToArrayAsync()
-                .ConfigureAwait(false);
-
-            var wordsSets = include.Select(g =>
-            {
-                var description = g.CreationDescriptions.Single(d => d.LanguageId == 2);
-                return new WordSetInfo()
-                {
-                    Id = g.GameId,
-                    Picture = g.Picture,
-                    Title = description.Title,
-                    Description = description.Description
-                };
-            });
-            wordSetReply.WordSets.AddRange(wordsSets);
-
-            return wordSetReply;
-        }
     }
 }