+++ WebApi/InWords.WebApi/InWords.WebApi/releasenotes.md
@@ -1,3 +1,14 @@
-﻿* add ready to run compilation
-* update webapi pipeline
-* add automatic git release
+﻿### Release notes
+
+- [x] INW-147 fix history dubles
+- [x] INW-148 remove redundant pipeline artefacts
+- [x] INW-149 update default router
+
+### NEW experimental API
+- [ ] INW-132 all levels request
+
+* GRPC WordSets.v2.GetFullsets
+
+* REST GET WordSets.Fullsets
+
+Load all games with stars using pagination.